[TASK] Cleanup example code
The tx_svconnector_base class contains sample
code which creates errors in IDEs, because it is imcomplete code.
It was meant as an example, but is not really useful anymore
since there are many real connector services available now.

It is cleaner to provide a useful phpDoc block and make the
various methods abstract to force implementation.

Resolves: #52100
Releases: 2.5
Change-Id: Ica57bf8eb5b60576b2c7faec1ab4522f6a323919
Reviewed-on: https://review.typo3.org/29132
Reviewed-by: Francois Suter
Tested-by: Francois Suter